Move bulk of Form View to take Date from MovieFormViewModel. Changed types to nullable and Stock to byte instead of int

+++ Vidly/Models/Movie.cs
@@ -28,6 +28,6 @@ namespace Vidly.Models
         public DateTime DateAdded { get; set; }
 
         [Display(Name = "Number in Stock")]
-        public int Stock { get; set; }
+        public byte Stock { get; set; }
     }
 }
